move the Api definition some easier to spot
